Move activity nav logic to top nav reducer
