+++ src/reducers/navigation/__tests__/top-test.js
@@ -415,4 +415,30 @@ describe('src/reducers/navigation/top', () => {
         .toEqual(createInitialState());
     });
   });
+
+  describe('CATEGORY_CHOOSE', () => {
+    it('should push on the category route', () => {
+      expect(reduce(createStack(), activities.chooseCategory(23)))
+        .toEqual(push(createStack(), createRoute(routes.ROUTE_CATEGORY, {
+          categoryId: 23,
+        })));
+    });
+  });
+
+  describe('ACTIVITY_CHOOSE', () => {
+    it('should push on the activity route', () => {
+      expect(reduce(createStack(), activities.chooseActivity(23)))
+        .toEqual(push(createStack(), createRoute(routes.ROUTE_ACTIVITY, {
+          activityId: 23,
+        })));
+    });
+  });
+
+  describe('ACTIVITY_SCREEN_DISMISS', () => {
+    it('should pop the current route', () => {
+      const stack = push(createStack(), createStack(routes.ROUTE_ACTIVITY));
+      expect(reduce(stack, activities.dismissActivityScreen()))
+        .toEqual(pop(stack));
+    });
+  });
 });
